# swagger/swagger.yml
basePath: /v1
consumes:
- application/json
definitions:
  Metadata_properties:
    properties:
      dc:description:
        type: string
      dc:language:
        type: string
      dc:publisher:
        type: string
      dc:rights:
        type: string
      dc:source:
        type: string
      dc:title:
        type: string
      nxtag:tags:
        items:
          $ref: '#/definitions/nuxeo_tags'
        type: array
    type: object
  Nuxeo_dublin_core_metadata:
    properties:
      properties:
        $ref: '#/definitions/Metadata_properties'
    type: object
...
  nuxeo_tags:
    properties:
      label:
        type: string
      username:
        type: string
    type: object
...
info:
  description: Api mid para la autenticacion de documentos en Nuxeo
  title: gestor_documental_mid
  version: '1.0'
paths:
...
  /document/{uid}/metadata:
    parameters:
    - description: UID del documento generado en Nuxeo
      in: path
      name: uid
      required: true
      type: string
    post:
      description: "Parameters\n----------\nrequest : uid, json\n    uid del documento\n\
        \    Json Body {Metadata}, metadata\n\nReturns\n-------\nResponse\n    Respuesta\
        \ con cuerpo, status y en formato json"
      operationId: post_document_post_metadata_/document/<string:uid>/metadata
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/Nuxeo_dublin_core_metadata'
      responses:
        '200':
          description: Success
        '400':
          description: Bad request
        '404':
          description: Not found
        '500':
          description: Nuxeo error
      summary: Permite cargar metadata relacionada a documento
      tags:
      - document
produces:
- application/json
...
swagger: '2.0'
...